requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
llama-index==0.11.16
llama-parse==0.5.7
python-dotenv==1.0.1
requests==2.32.3
Flask-Login==0.6.3
azure-storage-blob==12.23.1
arize-phoenix-otel==0.6.1
openinference-instrumentation-llama_index==3.0.4
openinference-instrumentation-openai==0.1.17